+++ app/src/main/java/woowacourse/movie/data/MockScreeningRepository.kt
@@ -1,7 +1,7 @@
 package woowacourse.movie.data
 
 import woowacourse.movie.R
-import woowacourse.movie.domain.screening.Movie
+import woowacourse.movie.domain.Movie
 import woowacourse.movie.domain.screening.Screening
 import java.time.LocalDate
